feat(stella-cli): the two-phase witness, and two of D5's silent surfaces recall again (#3243 Phase 3)
The epic's witness: a turn that opens outside a skill's domain and drifts
into it receives that skill by re-query, selected against touched paths —
structurally unreachable on base, where every selector fires once against
the opening prompt. Fleet workers and resumed runs stop passing
NoContextRecall: both now open the workspace's SessionMemory as their
ContextRecallPort, degrading to no frames exactly as before when there is
no memory. fleet_cmd.rs crossed the ratchet at the wiring and split its
inline tests to fleet_cmd/tests.rs, per the first-crossing rule.

Refs #3243
